refactor: split settings module
